backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-core==2.27.2
aiofiles==24.1.0
httpx==0.28.1
openai==1.58.1
python-multipart==0.0.20
annotated-types==0.7.0
